refactor(disbursements): nest NewDisbursement under AccountsPayables
Move the gem's A/P disbursement payload class from the top-level
`Agris::Api` namespace into `Agris::Api::AccountsPayables`, alongside
the `Disbursements` module that consumes it. Aligns with the broader
shift away from flat `lib/agris/api/*.rb` files toward organized
domain folders (mirroring `Grain::*`, `Inventory::*`, etc.).

Renames + content changes:
- `lib/agris/api/new_disbursement.rb`
  → `lib/agris/api/accounts_payables/new_disbursement.rb`
- `spec/lib/agris/api/new_disbursement_spec.rb`
  → `spec/lib/agris/api/accounts_payables/new_disbursement_spec.rb`
- Class is now `Agris::Api::AccountsPayables::NewDisbursement`.
- `lib/agris/api.rb` drops its top-level `NewDisbursement` autoload.
- `lib/agris/api/accounts_payables.rb` gains a nested
  `autoload :NewDisbursement` entry next to `:Disbursements`.

Breaking for downstream callers: `Agris::Api::NewDisbursement` is no
longer defined. Consumers (vflow_on_rails `TranslateInvoiceToDisbursement`
is the only known one) need to reference
`Agris::Api::AccountsPayables::NewDisbursement`. No backwards-compatible
alias is added — this PR has not yet shipped, so no public consumers
depend on the old path.

Behavior is unchanged: same 20 ACPD0 attributes, same `record_type`
default, same `to_xml_hash` wire output. Full suite 46/46 green on
Ruby 3.4.8.

Spec: docs/specs/SPEC-016-disbursement-pipeline.md
Task: VF-247
AC: AC-1, AC-2, AC-3, AC-4
