+++ lib/agris/api.rb
@@ -9,7 +9,6 @@ module Agris
     autoload :Inventory, 'agris/api/inventory'
     autoload :Messages, 'agris/api/messages'
     autoload :Name, 'agris/api/name'
-    autoload :NewDisbursement, 'agris/api/new_disbursement'
     autoload :NewOrder, 'agris/api/new_order'
     autoload :NewOrderRemark, 'agris/api/new_order_remark'
     autoload :NewVoucher, 'agris/api/new_voucher'
